# apps/api/openapi/moduflow-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "summary": "Create a user in the active organization",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List users in the active organization",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "summary": "Get a user by id in the active organization",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "summary": "Update a user in the active organization",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a user from the active organization",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Get a role by id in the active organization",
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Update a role in the active organization",
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Delete a role from the active organization",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/users/{userId}/roles": {
      "post": {
        "operationId": "UserRolesController_assignRole",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleResponse"
                }
              }
            }
          }
        },
        "summary": "Assign a role to a user",
        "tags": [
          "User Roles"
        ]
      },
      "get": {
        "operationId": "UserRolesController_findAll",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRoleResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List roles assigned to a user",
        "tags": [
          "User Roles"
        ]
      }
    },
    "/api/users/{userId}/roles/{roleId}": {
      "delete": {
        "operationId": "UserRolesController_removeRole",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleResponse"
                }
              }
            }
          }
        },
        "summary": "Remove a role from a user",
        "tags": [
          "User Roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moduflow API",
    "description": "Moduflow API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Unique email address for the user.",
            "example": "alex@example.com"
          },
          "name": {
            "type": "string",
            "description": "Display name for the user.",
            "example": "Alex Johnson",
            "maxLength": 120
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar image URL.",
            "example": "https://example.com/avatar.png",
            "maxLength": 2048
          }
        },
        "required": [
          "email"
        ]
      },
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Northwind Homes"
          },
          "slug": {
            "type": "string",
            "example": "northwind-homes"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "RoleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Admin"
          },
          "description": {
            "type": "object",
            "example": "Can manage users and permissions.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "email": {
            "type": "string",
            "example": "alex@example.com"
          },
          "name": {
            "type": "object",
            "example": "Alex Johnson",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "example": "https://example.com/avatar.png",
            "nullable": true
          },
          "organization": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrganizationResponse"
              }
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "avatarUrl",
          "organization",
          "roles"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Unique email address for the user.",
            "example": "alex@example.com"
          },
          "name": {
            "type": "string",
            "description": "Display name for the user.",
            "example": "Alex Johnson",
            "maxLength": 120
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar image URL.",
            "example": "https://example.com/avatar.png",
            "maxLength": 2048
          }
        }
      },
...
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Role name.",
            "example": "Admin",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Role description.",
            "example": "Can manage users and permissions."
          }
        }
      },
...
      "AssignUserRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "description": "Role id to assign to the user.",
            "example": "1"
          }
        },
        "required": [
          "roleId"
        ]
      },
      "UserRoleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Admin"
          },
          "description": {
            "type": "object",
            "example": "Can manage users and permissions.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
...
    }
  }
}