# apps/api/src/user-roles/user-roles.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Post,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiOkResponse,
  ApiOperation,
  ApiTags,
} from '@nestjs/swagger';
import { plainToInstance } from 'class-transformer';
import { parseBigIntId } from '../common/ids/parse-bigint-id';
import { OrganizationId } from '../platform/organization-id.decorator';
import { platformPath } from '../platform/platform.constants';
import { PlatformOrganizationContextGuard } from '../platform/platform-organization-context.guard';
import {
  ApiBigIntIdParam,
  ApiOrganizationHeader,
} from '../platform/platform-swagger.decorator';
import { AssignUserRoleDto } from './dto/assign-user-role.dto';
import { UserRoleResponse } from './dto/user-role.dto';
import { UserRolesService } from './user-roles.service';

@ApiTags('User Roles')
@ApiOrganizationHeader()
@UseGuards(PlatformOrganizationContextGuard)
@Controller(platformPath('users/:userId/roles'))
export class UserRolesController {
  constructor(private readonly userRolesService: UserRolesService) {}

  @Post()
  @ApiOperation({ summary: 'Assign a role to a user' })
  @ApiBigIntIdParam('userId', 'user')
  @ApiBody({ type: AssignUserRoleDto })
  @ApiCreatedResponse({ type: UserRoleResponse })
  async assignRole(
    @OrganizationId() organizationId: bigint,
    @Param('userId') userId: string,
    @Body('roleId') roleId: string,
  ) {
    const role = await this.userRolesService.assignRole(
      organizationId,
      parseBigIntId(userId, 'userId'),
      parseBigIntId(roleId, 'roleId'),
    );
    return plainToInstance(UserRoleResponse, role, {
      excludeExtraneousValues: true,
    });
  }

  @Get()
  @ApiOperation({ summary: 'List roles assigned to a user' })
  @ApiBigIntIdParam('userId', 'user')
  @ApiOkResponse({ type: UserRoleResponse, isArray: true })
  async findAll(
    @OrganizationId() organizationId: bigint,
    @Param('userId') userId: string,
  ) {
    const roles = await this.userRolesService.findAll(
      organizationId,
      parseBigIntId(userId, 'userId'),
    );
    return plainToInstance(UserRoleResponse, roles, {
      excludeExtraneousValues: true,
    });
  }

  @Delete(':roleId')
  @ApiOperation({ summary: 'Remove a role from a user' })
  @ApiBigIntIdParam('userId', 'user')
  @ApiBigIntIdParam('roleId', 'role')
  @ApiOkResponse({ type: UserRoleResponse })
  async removeRole(
    @OrganizationId() organizationId: bigint,
    @Param('userId') userId: string,
    @Param('roleId') roleId: string,
  ) {
    const role = await this.userRolesService.removeRole(
      organizationId,
      parseBigIntId(userId, 'userId'),
      parseBigIntId(roleId, 'roleId'),
    );
    return plainToInstance(UserRoleResponse, role, {
      excludeExtraneousValues: true,
    });
  }
}